requirements.txt
streamlit
azure-cognitiveservices-speech
python-dotenv